Extract fttc
