+++ helpers/index.js
@@ -57,7 +57,8 @@ async function fetchCombos(challenge, reload=false) {
 }
 
 const { scrapeAll2TCCombos } = require('../services/index/2tc_scraper.js')
-const { scrapeAll2MPCompletions } = require('../services/index/2mp_scraper.js')
+const { scrapeAll2MPCompletions } = require('../services/index/2mp_scraper.js');
+const { scrapeAllFTTCCombos } = require('../services/index/fttc_scraper');
 
 async function scrapeAllCombos(challenge) {
     switch(challenge) {
@@ -65,6 +66,8 @@ async function scrapeAllCombos(challenge) {
             return await scrapeAll2TCCombos();
         case '2mp':
             return await scrapeAll2MPCompletions();
+        case 'fttc':
+            return await scrapeAllFTTCCombos();
         default:
             throw 'Challenge not found'
     }
